removed tailwind and created separate folder for each component and added style.css file for each component
