+++ src/Components/About/About.jsx
@@ -1,3 +1,4 @@
+import './style.css';
 function About() {
   return (
     <div>This is About</div>
