+++ src/Components/Blog/Blog.jsx
@@ -1,3 +1,4 @@
+import './style.css';
 function Blog() {
   return (
     <div>This is Blog</div>
