+++ src/Components/Footer/Footer.jsx
@@ -1,3 +1,4 @@
+import './style.css';
 function Footer() {
   return (
     <div>This is Footer</div>
