+++ src/Components/Navbar/Navbar.jsx
@@ -1,10 +1,11 @@
+import './style.css';
 function Navbar() {
   const handleOnClick = (bodyElementID) => {
     document.body.querySelector(`[id=${bodyElementID}]`)?.scrollIntoView({ behavior: 'smooth', block: 'start', inline: 'nearest' });
   }
   return (
     <>
-      <nav className={'navbar navbar-expand-lg navbar-dark bg-dark tw-bg-gradient-to-r tw-from-sky-900 tw-via-sky-900 tw-to-sky-900'}>
+      <nav className={'navbar navbar-expand-lg navbar-light bg-light bglight'}>
         <div className='container-fluid'>
           <img src="https://www.w3schools.com/images/lamp.jpg" alt="Lamp" width="32" height="32"></img> {/*company logo here*/}
           <button className='navbar-toggler' type='button' data-bs-toggle='collapse' data-bs-target='#navbarSupportedContent' aria-controls='navbarSupportedContent' aria-expanded='false' aria-label='Toggle navigation'>
@@ -13,13 +14,13 @@ function Navbar() {
           <div className='collapse navbar-collapse justify-content-end' id='navbarSupportedContent'>
             <ul className='navbar-nav'>
               <li className='nav-item'>
-                <button className={'nav-link !tw-text-yellow-300'} onClick={() => handleOnClick('about')}>About</button>
+                <button className={'nav-link navtitleColor'} onClick={() => handleOnClick('about')}>About</button>
               </li>
               <li className='nav-item'>
-                <button className={'nav-link !tw-text-yellow-300'} onClick={() => handleOnClick('blog')}>Blog</button>
+                <button className={'nav-link navtitleColor'} onClick={() => handleOnClick('blog')}>Blog</button>
               </li>
               <li className='nav-item'>
-                <button className={'nav-link !tw-text-yellow-300'} onClick={() => handleOnClick('services')}>Services</button>
+                <button className={'nav-link navtitleColor'} onClick={() => handleOnClick('services')}>Services</button>
               </li>
             </ul>
           </div>
