+++ src/Components/Navbar/style.css
@@ -1,0 +1,11 @@
+.bglight{
+    background:linear-gradient(rgb(0, 110, 177) 0%, rgb(0, 110, 177) 50%);
+}
+
+.linktag{
+    text-decoration: none;
+}
+
+.navtitleColor{
+    color: yellow !important;
+}
