+++ src/Components/Services/Services.jsx
@@ -1,3 +1,4 @@
+import './style.css';
 function Services() {
   return (
     <div>This is Services</div>
